starting test for character

+++ src/Game/SaWo/Property/SaWoTrait.php
@@ -52,7 +52,7 @@ abstract class SaWoTrait implements Property {
     }
 
     public function getDifficulty() {
-        return $this->dice / 2 + 2 + floor($this->bonus / 2);
+        return (int) ($this->dice / 2 + 2 + floor($this->bonus / 2));
     }
 
     public function __toString() {
